backend/requirements.txt
asttokens==3.0.0
django==5.1.6
python-decouple==3.8
firebase-admin==6.9.0
openpyxl==3.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
celery==5.5.3
weasyprint==65.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
psycopg2==2.9.10
psycopg2-binary==2.9.10